Do not configure logwatch.conf if ensure absent
If someone is using the module to ensure the package is absent, first
you might be asking, why use the module, but it makes sense in a large
environment that some systems may not need logwatch.

This makes sure that if ensure_package => absent then the logwatch.conf
does not try to get managed / modified which would fail.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>